requirements.txt
pillow==11.1.0
numpy==2.2.4
scipy==1.16.0
tkinterdnd2==0.4.3